+++ src/core/search/analysis/index.ts
@@ -1,4 +1,5 @@
 export * from "./channels.js";
+export * from "./budget.js";
 export * from "./fields.js";
 export * from "./korean.js";
 export * from "./query.js";